fix: resolve clippy warnings and formatting issues
Fix never_loop in CLI, add Default impls for LcpState and IpcpState,
replace match-for-equality with if/if-let, remove unnecessary casts
and return statements, use std::io::Error::other.

+++ src-tauri/crates/fortivpn/src/tun.rs
@@ -8,7 +8,7 @@ pub fn create_tun(ip: Ipv4Addr, peer_ip: Ipv4Addr, mtu: u16) -> Result<AsyncDevi
     let mut config = Configuration::default();
     config.address(ip);
     config.destination(peer_ip);
-    config.mtu(mtu as u16);
+    config.mtu(mtu);
     config.up();
 
     #[cfg(target_os = "macos")]
